+++ tensorboard/webapp/plugins/text_v2/data_source/BUILD
@@ -9,7 +9,7 @@ ng_module(
         "index.ts",
         "text_v2_data_source.ts",
         "text_v2_data_source_module.ts",
-        "text_v2_types.ts",
+        "text_v2_server_data_source.ts",
     ],
     deps = [
         "//tensorboard/webapp/webapp_data_source:http_client",
