+++ tensorboard/webapp/plugins/text_v2/data_source/index.ts
@@ -12,4 +12,4 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
-export * from './text_v2_types';
+export * from './text_v2_data_source';
