--- tensorboard/webapp/plugins/text_v2/data_source/text_v2_types.ts
@@ -1,42 +1,0 @@
-/* Copyright 2020 The TensorFlow Authors. All Rights Reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-==============================================================================*/
-import {Injectable} from '@angular/core';
-import {Observable} from 'rxjs';
-
-export interface BackendRunToTagsMap {
-  [runName: string]: string[];
-}
-
-export interface BackendStepDatum {
-  original_shape: number[];
-  step: number;
-  string_array: string[][];
-  wall_time: number;
-  truncated: boolean;
-}
-
-export interface StepDatum {
-  originalShape: number[];
-  step: number;
-  stringArray: string[][];
-  wallTimeInMs: number;
-  truncated: boolean;
-}
-
-@Injectable()
-export abstract class TextDataSource {
-  abstract fetchRunToTag(): Observable<BackendRunToTagsMap>;
-  abstract fetchTextData(run: string, tag: string): Observable<StepDatum[]>;
-}
